Massive style fix (braces, indentation, etc.)

+++ crawl-ref/source/spl-clouds.cc
@@ -237,9 +237,11 @@ void manage_fire_shield(int delay)
         you.duration[DUR_FIRE_SHIELD] = 0;
 
     // Remove fire clouds on top of you
-    if (env.cgrid(you.pos()) != EMPTY_CLOUD)
-        if (env.cloud[env.cgrid(you.pos())].type == CLOUD_FIRE)
-            delete_cloud_at(you.pos());
+    if (env.cgrid(you.pos()) != EMPTY_CLOUD
+        && env.cloud[env.cgrid(you.pos())].type == CLOUD_FIRE)
+    {
+        delete_cloud_at(you.pos());
+    }
 
     if (!you.duration[DUR_FIRE_SHIELD])
     {
